+++ docs/en/changes/changes.md
@@ -62,6 +62,7 @@
 
 * Bump up netty to 4.2.4.Final.
 * Bump up commons-lang to 3.18.0.
+* BanyanDB: support group `replicas` and `user/password` for basic authentication.
 
 #### UI